ScrollView + Drag Gesture problem fixed
