+++ Sources/Internal/Managers/ScreenManager.swift
@@ -70,26 +70,3 @@ class ScreenManager: ObservableObject {
     private init() {}
 }
 #endif
-
-
-// MARK: - Common Implementation
-extension View {
-    func updateScreenSize() -> some View {
-        GeometryReader { reader in
-            onAppear { ScreenManager.update(reader) }
-                .onChange(of: reader.size) { _ in ScreenManager.update(reader) }
-                .onChange(of: reader.safeAreaInsets) { _ in ScreenManager.update(reader) }
-        }
-    }
-}
-fileprivate extension ScreenManager {
-    static func update(_ reader: GeometryProxy) {
-        shared.size.height = reader.size.height + reader.safeAreaInsets.top + reader.safeAreaInsets.bottom
-        shared.size.width = reader.size.width + reader.safeAreaInsets.leading + reader.safeAreaInsets.trailing
-
-        shared.safeArea.top = reader.safeAreaInsets.top
-        shared.safeArea.bottom = reader.safeAreaInsets.bottom
-        shared.safeArea.left = reader.safeAreaInsets.leading
-        shared.safeArea.right = reader.safeAreaInsets.trailing
-    }
-}
